Moved most of the PETSc implementation over to cpp files.
