+++ include/neso_particles/external_interfaces/petsc/dmplex_helper.hpp
@@ -21,18 +21,8 @@ constexpr static char face_sets_label[] = "Face Sets";
  * @param[in] comm MPI communicator, default MPI_COMM_WORLD.
  * @param[in] overlap Optional overlap to pass to PETSc (default 0).
  */
-inline void generic_distribute(DM *dm, MPI_Comm comm = MPI_COMM_WORLD,
-                               const PetscInt overlap = 0) {
-  int size;
-  MPICHK(MPI_Comm_size(comm, &size));
-  if (size > 1) {
-    DM dm_out;
-    PETSCCHK(DMPlexDistribute(*dm, overlap, nullptr, &dm_out));
-    NESOASSERT(dm_out, "Could not distribute mesh.");
-    PETSCCHK(DMDestroy(dm));
-    *dm = dm_out;
-  }
-}
+void generic_distribute(DM *dm, MPI_Comm comm = MPI_COMM_WORLD,
+                        const PetscInt overlap = 0);
 
 /**
  * Setup the coordinate section for a DMPlex. See
@@ -42,21 +32,8 @@ inline void generic_distribute(DM *dm, MPI_Comm comm = MPI_COMM_WORLD,
  * @param vertex_start Coordinate index for first coordinate.
  * @param vertex_end Coordinate index +1 for last coordinate.
  */
-inline void setup_coordinate_section(DM &dm, const PetscInt vertex_start,
-                                     const PetscInt vertex_end) {
-  PetscInt ndim;
-  PETSCCHK(DMGetCoordinateDim(dm, &ndim));
-  PetscSection coord_section;
-  PETSCCHK(DMGetCoordinateSection(dm, &coord_section));
-  PETSCCHK(PetscSectionSetNumFields(coord_section, 1));
-  PETSCCHK(PetscSectionSetFieldComponents(coord_section, 0, ndim));
-  PETSCCHK(PetscSectionSetChart(coord_section, vertex_start, vertex_end));
-  for (PetscInt v = vertex_start; v < vertex_end; ++v) {
-    PETSCCHK(PetscSectionSetDof(coord_section, v, ndim));
-    PETSCCHK(PetscSectionSetFieldDof(coord_section, v, 0, ndim));
-  }
-  PETSCCHK(PetscSectionSetUp(coord_section));
-}
+void setup_coordinate_section(DM &dm, const PetscInt vertex_start,
+                              const PetscInt vertex_end);
 
 /**
  * Setup a PETSc vector in which coordinates can be get/set for local mesh. See
@@ -66,21 +43,7 @@ inline void setup_coordinate_section(DM &dm, const PetscInt vertex_start,
  * @param[in, out] coordinates Vector to setup to use with coordinates.
  * VecDestroy should be called on this vector.
  */
-inline void setup_local_coordinate_vector(DM &dm, Vec &coordinates) {
-  PetscSection coord_section;
-  PETSCCHK(DMGetCoordinateSection(dm, &coord_section));
-
-  // create the actual coordinates vector
-  PetscInt coord_size;
-  PetscInt ndim;
-  PETSCCHK(DMGetCoordinateDim(dm, &ndim));
-  PETSCCHK(PetscSectionGetStorageSize(coord_section, &coord_size));
-  PETSCCHK(VecCreate(PETSC_COMM_SELF, &coordinates));
-  PETSCCHK(PetscObjectSetName((PetscObject)coordinates, "coordinates"));
-  PETSCCHK(VecSetSizes(coordinates, coord_size, PETSC_DETERMINE));
-  PETSCCHK(VecSetBlockSize(coordinates, ndim));
-  PETSCCHK(VecSetType(coordinates, VECSTANDARD));
-}
+void setup_local_coordinate_vector(DM &dm, Vec &coordinates);
 
 /**
  * Class to determine new local indices from global indices when constructing
@@ -199,120 +162,10 @@ struct HaloDMIndexMapper {
  * rank, global petsc index of cell}.
  * @returns True if the constructed DMPlex is not empty otherwise false.
  */
-inline bool
-dm_from_serialised_cells(std::list<DMPlexCellSerialise> &serialised_cells,
-                         DM &dm_prototype, DM &dm,
-                         std::map<PetscInt, std::tuple<int, PetscInt, PetscInt>>
-                             &map_local_lid_remote_lid) {
-
-  const PetscInt num_cells = serialised_cells.size();
-  std::vector<CellSTDRepresentation> std_rep_cells(num_cells);
-  int index = 0;
-  for (auto &sc : serialised_cells) {
-    std_rep_cells.at(index).deserialise(sc.cell_representation);
-    index++;
-  }
-
-  HaloDMIndexMapper index_mapper(std_rep_cells);
-  // create the map from the new local dm_halo indices to the original local
-  // indices on the remote ranks
-
-  for (auto &sc : serialised_cells) {
-    const auto global_index = sc.cell_global_id;
-    const PetscInt remote_local_id = sc.cell_local_id;
-    const int remote_rank = sc.owning_rank;
-    const auto local_local_id =
-        index_mapper.get_local_point_index(global_index);
-    map_local_lid_remote_lid[local_local_id] = {remote_rank, remote_local_id,
-                                                global_index};
-  }
-
-  // Create the new DMPlex.
-  PETSCCHK(DMCreate(PETSC_COMM_SELF, &dm));
-  PETSCCHK(DMSetType(dm, DMPLEX));
-
-  if (num_cells > 0) {
-
-    PetscInt tmp_int;
-    PETSCCHK(DMGetDimension(dm_prototype, &tmp_int));
-    PETSCCHK(DMSetDimension(dm, tmp_int));
-    PETSCCHK(DMGetCoordinateDim(dm_prototype, &tmp_int));
-    const PetscInt ndim_coord = tmp_int;
-    PETSCCHK(DMSetCoordinateDim(dm, tmp_int));
-
-    PETSCCHK(
-        DMPlexSetChart(dm, index_mapper.chart_start, index_mapper.chart_end));
-
-    std::set<PetscInt> points_set;
-    for (auto &std_cell : std_rep_cells) {
-      for (auto &point_spec : std_cell.point_specs) {
-        const PetscInt global_point = point_spec.first;
-        if (!points_set.count(global_point)) {
-          points_set.insert(global_point);
-          const PetscInt local_point =
-              index_mapper.get_local_point_index(global_point);
-          const PetscInt cone_size = point_spec.second.size();
-          PETSCCHK(DMPlexSetConeSize(dm, local_point, cone_size));
-        }
-      }
-    }
-
-    PETSCCHK(DMSetUp(dm));
-    std::vector<PetscInt> cone_local;
-    points_set.clear();
-    for (auto &std_cell : std_rep_cells) {
-      for (auto &point_spec : std_cell.point_specs) {
-        const PetscInt global_point = point_spec.first;
-        if (!points_set.count(global_point)) {
-          points_set.insert(global_point);
-          const PetscInt local_point =
-              index_mapper.get_local_point_index(global_point);
-          auto &cone_global = point_spec.second;
-          cone_local.clear();
-          cone_local.reserve(cone_global.size());
-          for (auto gx : cone_global) {
-            cone_local.push_back(index_mapper.get_local_point_index(gx));
-          }
-          PETSCCHK(DMPlexSetCone(dm, local_point, cone_local.data()));
-          if (!map_local_lid_remote_lid.count(local_point)) {
-            map_local_lid_remote_lid[local_point] = {-1, -1, global_point};
-          }
-        }
-      }
-    }
-
-    PETSCCHK(DMPlexSymmetrize(dm));
-    PETSCCHK(DMPlexStratify(dm));
-
-    PetscInt vertex_start, vertex_end;
-    index_mapper.get_depth_stratum(0, &vertex_start, &vertex_end);
-    setup_coordinate_section(dm, vertex_start, vertex_end);
-    Vec coordinates;
-    PetscScalar *coords;
-    PetscInterface::setup_local_coordinate_vector(dm, coordinates);
-    PETSCCHK(VecGetArray(coordinates, &coords));
-
-    // write coordinates of vertices
-    for (auto &std_cell : std_rep_cells) {
-      for (const auto &point_vertex : std_cell.vertices) {
-        const PetscInt global_point = point_vertex.first;
-        const PetscInt local_point =
-            index_mapper.get_local_point_index(global_point);
-        const PetscInt vertex_index = local_point - vertex_start;
-        for (PetscInt dimx = 0; dimx < ndim_coord; dimx++) {
-          const PetscScalar value = point_vertex.second.at(dimx);
-          coords[vertex_index * ndim_coord + dimx] = value;
-        }
-      }
-    }
-
-    PETSCCHK(VecRestoreArray(coordinates, &coords));
-    PETSCCHK(DMSetCoordinatesLocal(dm, coordinates));
-    PETSCCHK(VecDestroy(&coordinates));
-  }
-
-  return num_cells > 0;
-}
+bool dm_from_serialised_cells(
+    std::list<DMPlexCellSerialise> &serialised_cells, DM &dm_prototype, DM &dm,
+    std::map<PetscInt, std::tuple<int, PetscInt, PetscInt>>
+        &map_local_lid_remote_lid);
 
 /**
  * Helper class that wraps a PETSc DMPlex and simplifies common operations.
@@ -368,35 +221,12 @@ public:
    * @param local_index Local index of the cell to get a representation of.
    * @returns Serialisable representation of the cell.
    */
-  inline DMPlexCellSerialise get_copyable_cell(const PetscInt local_index) {
-    this->check_valid_local_cell(local_index);
-    const PetscInt point = this->map_np_to_petsc.at(local_index);
-
-    int rank;
-    MPICHK(MPI_Comm_rank(this->comm, &rank));
-    DMPolytopeType cell_type;
-    PETSCCHK(DMPlexGetCellType(dm, point, &cell_type));
-
-    auto lambda_rename = [&](PetscInt cell) -> PetscInt {
-      return this->get_point_global_index(cell);
-    };
-    std::function<PetscInt(PetscInt)> rename_function = lambda_rename;
-    auto spec =
-        PetscInterface::CellSTDRepresentation(this->dm, point, rename_function);
-
-    std::vector<std::byte> cell_representation;
-    spec.serialise(cell_representation);
-
-    DMPlexCellSerialise cs{local_index, lambda_rename(point), rank, cell_type,
-                           cell_representation};
-
-    return cs;
-  }
+  DMPlexCellSerialise get_copyable_cell(const PetscInt local_index);
 
   /**
    * Free the helper. Must be called collectively on the communicator.
    */
-  inline void free() { PETSCCHK(ISDestroy(&this->global_point_numbers)); }
+  void free();
 
   /**
    * Construct helper class from DMPlex. Collective on the communicator.
@@ -404,40 +234,12 @@ public:
    * @param comm MPI communicator the DMPlex is constructed on.
    * @param dm Input DMPlex to wrap.
    */
-  DMPlexHelper(MPI_Comm comm, DM dm)
-      : volume(-1.0), bounding_box(nullptr), comm(comm), dm(dm) {
-    DMPlexInterpolatedFlag interpolated;
-    PETSCCHK(DMPlexIsInterpolated(this->dm, &interpolated));
-    NESOASSERT(interpolated == DMPLEX_INTERPOLATED_FULL,
-               "Expected fully interpolated mesh.");
-    PETSCCHK(DMGetCoordinateDim(this->dm, &this->ndim));
-    PETSCCHK(DMPlexGetHeightStratum(this->dm, 0, &this->cell_start,
-                                    &this->cell_end));
-    PETSCCHK(DMPlexGetChart(this->dm, &this->point_start, &this->point_end));
-    PETSCCHK(DMPlexCreatePointNumbering(this->dm, &this->global_point_numbers));
-
-    this->map_np_to_petsc.clear();
-    PetscInt ix = 0;
-    for (int cx = this->cell_start; cx < this->cell_end; cx++) {
-      auto global = this->internal_get_point_global_index(cx);
-      if (global > -1) {
-        this->map_np_to_petsc.push_back(cx);
-        this->map_petsc_to_np[cx] = ix++;
-      }
-    }
-
-    NESOASSERT(static_cast<std::size_t>(ix) == this->map_petsc_to_np.size(),
-               "Size missmatch.");
-    NESOASSERT(static_cast<std::size_t>(ix) == this->map_np_to_petsc.size(),
-               "Size missmatch.");
-    this->ncells = this->map_np_to_petsc.size();
-    NESOASSERT(this->ncells, "A rank has zero cells.");
-  }
+  DMPlexHelper(MPI_Comm comm, DM dm);
 
   /**
    * @returns the number of cells.
    */
-  inline int get_cell_count() { return this->ncells; }
+  int get_cell_count();
 
   /**
    * Convert a local cell id into a DMPlex cell id.
@@ -445,12 +247,7 @@ public:
    * @param local_index Local index in [0, num_cells).
    * @returns DMPlex point index in [cell_start, cell_end).
    */
-  inline PetscInt get_dmplex_cell_index(const PetscInt local_index) {
-    this->check_valid_local_cell(local_index);
-    const auto index = this->map_np_to_petsc.at(local_index);
-    this->check_valid_petsc_cell(index);
-    return index;
-  }
+  PetscInt get_dmplex_cell_index(const PetscInt local_index);
 
   /**
    * Convert a DMPlex cell id into a local cell id.
@@ -458,12 +255,7 @@ public:
    * @param petsc_index Local index in [cell_start, cell_end).
    * @returns Local point index in [0, num_cells).
    */
-  inline PetscInt get_local_cell_index(const PetscInt petsc_index) {
-    this->check_valid_petsc_cell(petsc_index);
-    const auto index = this->map_petsc_to_np.at(petsc_index);
-    this->check_valid_local_cell(index);
-    return index;
-  }
+  PetscInt get_local_cell_index(const PetscInt petsc_index);
 
   /**
    * Remove the negation from point indices which DMPlex uses to denote global
@@ -472,9 +264,7 @@ public:
    * @param c Input index.
    * @returns c if c > -1 else ((c * (-1)) - 1)
    */
-  inline PetscInt signed_global_id_to_global_id(const PetscInt c) {
-    return (c > -1) ? c : ((c * (-1)) - 1);
-  }
+  PetscInt signed_global_id_to_global_id(const PetscInt c);
 
   /**
    * Get the global index of a point from the local point index.
@@ -484,36 +274,15 @@ public:
    * index.
    * @returns Global point index.
    */
-  inline PetscInt get_point_global_index(const PetscInt point,
-                                         const bool signed_point = false) {
-    NESOASSERT((this->point_start <= point) && (point < this->point_end),
-               "Bad point passed.");
-    PetscInt global_point = this->internal_get_point_global_index(point);
-
-    if (signed_point) {
-      return global_point;
-    } else {
-      return signed_global_id_to_global_id(global_point);
-    }
-  }
+  PetscInt get_point_global_index(const PetscInt point,
+                                  const bool signed_point = false);
 
   /**
    * Get a bounding box for the cells on this MPI rank.
    *
    * @returns Bounding box for cells in DMPlex.
    */
-  inline ExternalCommon::BoundingBoxSharedPtr get_bounding_box() {
-    // Create the bounding box on first use.
-    if (!this->bounding_box) {
-      auto bb = std::make_shared<ExternalCommon::BoundingBox>();
-      const auto num_cells = this->get_cell_count();
-      for (int cellx = 0; cellx < num_cells; cellx++) {
-        bb->expand(this->get_cell_bounding_box(cellx));
-      }
-      this->bounding_box = bb;
-    }
-    return this->bounding_box;
-  }
+  ExternalCommon::BoundingBoxSharedPtr get_bounding_box();
 
   /**
    * Get a bounding box for a mesh cell (assumes linear mesh).
@@ -521,43 +290,8 @@ public:
    * @param cell Local cell index.
    * @returns Bounding box for cell.
    */
-  inline ExternalCommon::BoundingBoxSharedPtr
-  get_cell_bounding_box(const PetscInt cell) {
-    this->check_valid_local_cell(cell);
-
-    std::vector<REAL> bb = {std::numeric_limits<REAL>::max(),
-                            std::numeric_limits<REAL>::max(),
-                            std::numeric_limits<REAL>::max(),
-                            std::numeric_limits<REAL>::lowest(),
-                            std::numeric_limits<REAL>::lowest(),
-                            std::numeric_limits<REAL>::lowest()};
-
-    for (int dx = this->ndim; dx < 3; dx++) {
-      bb[dx] = 0.0;
-      bb[dx + 3] = 0.0;
-    }
-
-    const PetscScalar *array;
-    PetscScalar *coords = nullptr;
-    PetscInt num_coords;
-    PetscBool is_dg;
-    const PetscInt petsc_index = this->map_np_to_petsc.at(cell);
-    this->check_valid_petsc_cell(petsc_index);
-    PETSCCHK(DMPlexGetCellCoordinates(dm, petsc_index, &is_dg, &num_coords,
-                                      &array, &coords));
-    NESOASSERT(coords != nullptr, "No vertices returned for cell.");
-    const PetscInt num_verts = num_coords / ndim;
-    for (PetscInt vx = 0; vx < num_verts; vx++) {
-      for (PetscInt dimx = 0; dimx < this->ndim; dimx++) {
-        const REAL cx = coords[vx * ndim + dimx];
-        bb[dimx] = std::min(bb[dimx], cx);
-        bb[dimx + 3] = std::max(bb[dimx + 3], cx);
-      }
-    }
-    PETSCCHK(DMPlexRestoreCellCoordinates(dm, petsc_index, &is_dg, &num_coords,
-                                          &array, &coords));
-    return std::make_shared<ExternalCommon::BoundingBox>(bb);
-  }
+  ExternalCommon::BoundingBoxSharedPtr
+  get_cell_bounding_box(const PetscInt cell);
 
   /**
    * Get the vertices of an edge using a PETSc index.
@@ -565,32 +299,8 @@ public:
    * @param[in] petsc_index PETSc point index.
    * @param[in, out] vertices Vector of vertices.
    */
-  inline void get_generic_vertices(const PetscInt petsc_index,
-                                   std::vector<std::vector<REAL>> &vertices) {
-
-    const PetscScalar *array;
-    PetscScalar *coords = nullptr;
-    PetscInt num_coords;
-    PetscBool is_dg;
-    this->check_valid_petsc_point(petsc_index);
-    PETSCCHK(DMPlexGetCellCoordinates(dm, petsc_index, &is_dg, &num_coords,
-                                      &array, &coords));
-    NESOASSERT(coords != nullptr, "No vertices returned for cell.");
-    const PetscInt num_verts = num_coords / ndim;
-
-    vertices.clear();
-    vertices.reserve(num_verts);
-    for (PetscInt vx = 0; vx < num_verts; vx++) {
-      std::vector<REAL> tmp(ndim);
-      for (PetscInt dimx = 0; dimx < this->ndim; dimx++) {
-        const REAL cx = coords[vx * ndim + dimx];
-        tmp.at(dimx) = cx;
-      }
-      vertices.push_back(tmp);
-    }
-    PETSCCHK(DMPlexRestoreCellCoordinates(dm, petsc_index, &is_dg, &num_coords,
-                                          &array, &coords));
-  }
+  void get_generic_vertices(const PetscInt petsc_index,
+                            std::vector<std::vector<REAL>> &vertices);
 
   /**
    * Get the vertices of a cell.
@@ -598,34 +308,8 @@ public:
    * @param[in] cell Local cell index.
    * @param[in, out] vertices Vector of vertices.
    */
-  inline void get_cell_vertices(const PetscInt cell,
-                                std::vector<std::vector<REAL>> &vertices) {
-    this->check_valid_local_cell(cell);
-
-    const PetscScalar *array;
-    PetscScalar *coords = nullptr;
-    PetscInt num_coords;
-    PetscBool is_dg;
-    const PetscInt petsc_index = this->map_np_to_petsc.at(cell);
-    this->check_valid_petsc_cell(petsc_index);
-    PETSCCHK(DMPlexGetCellCoordinates(dm, petsc_index, &is_dg, &num_coords,
-                                      &array, &coords));
-    NESOASSERT(coords != nullptr, "No vertices returned for cell.");
-    const PetscInt num_verts = num_coords / ndim;
-
-    vertices.clear();
-    vertices.reserve(num_verts);
-    for (PetscInt vx = 0; vx < num_verts; vx++) {
-      std::vector<REAL> tmp(ndim);
-      for (PetscInt dimx = 0; dimx < this->ndim; dimx++) {
-        const REAL cx = coords[vx * ndim + dimx];
-        tmp.at(dimx) = cx;
-      }
-      vertices.push_back(tmp);
-    }
-    PETSCCHK(DMPlexRestoreCellCoordinates(dm, petsc_index, &is_dg, &num_coords,
-                                          &array, &coords));
-  }
+  void get_cell_vertices(const PetscInt cell,
+                         std::vector<std::vector<REAL>> &vertices);
 
   /**
    * Get average of the vertices of a cell.
@@ -633,27 +317,7 @@ public:
    * @param[in] cell Local cell index.
    * @param[in, out] average Vector of average of vertices.
    */
-  inline void get_cell_vertex_average(const PetscInt cell,
-                                      std::vector<REAL> &average) {
-    this->check_valid_local_cell(cell);
-    NESOASSERT(average.size() == static_cast<std::size_t>(this->ndim),
-               "Missmatch between vector size and number of dimensions");
-
-    std::fill(average.begin(), average.end(), 0.0);
-    std::vector<std::vector<REAL>> vertices;
-    this->get_cell_vertices(cell, vertices);
-    const int num_verts = vertices.size();
-    for (PetscInt vx = 0; vx < num_verts; vx++) {
-      for (PetscInt dimx = 0; dimx < this->ndim; dimx++) {
-        const REAL cx = vertices.at(vx).at(dimx);
-        average.at(dimx) += cx;
-      }
-    }
-    const REAL tmp_factor = 1.0 / ((REAL)num_verts);
-    for (int dx = 0; dx < this->ndim; dx++) {
-      average.at(dx) *= tmp_factor;
-    }
-  }
+  void get_cell_vertex_average(const PetscInt cell, std::vector<REAL> &average);
 
   /**
    * Determine if mesh contains a point.
@@ -661,29 +325,7 @@ public:
    * @param[in] point Point to test.
    * @returns Negative value if point not located, otherwise owning cell.
    */
-  inline int contains_point(std::vector<PetscScalar> &point) {
-    const PetscInt ndim = this->ndim;
-    NESOASSERT(point.size() == static_cast<std::size_t>(ndim),
-               "Miss-match in point size and mesh dimension.");
-    Vec v;
-    PETSCCHK(VecCreate(MPI_COMM_SELF, &v));
-    PETSCCHK(VecSetSizes(v, ndim, ndim));
-    PETSCCHK(VecSetBlockSize(v, ndim));
-    PETSCCHK(VecSetFromOptions(v));
-    PetscScalar *v_ptr;
-    PETSCCHK(VecGetArrayWrite(v, &v_ptr));
-    for (int dimx = 0; dimx < ndim; dimx++) {
-      v_ptr[dimx] = point.at(dimx);
-    }
-    PETSCCHK(VecRestoreArrayWrite(v, &v_ptr));
-    PetscSF cell_sf = nullptr;
-    PETSCCHK(DMLocatePoints(dm, v, DM_POINTLOCATION_NONE, &cell_sf));
-    const PetscSFNode *cells;
-    PetscInt n_found;
-    const PetscInt *found;
-    PETSCCHK(PetscSFGetGraph(cell_sf, NULL, &n_found, &found, &cells));
-    return cells[0].index;
-  }
+  int contains_point(std::vector<PetscScalar> &point);
 
   /**
    * Determine if mesh cell contains a point.
@@ -692,113 +334,24 @@ public:
    * @param[in] point Point to test.
    * @returns true if point contains cell.
    */
-  inline bool cell_contains_point(const PetscInt index,
-                                  std::vector<PetscScalar> &point) {
-    const PetscInt ndim = this->ndim;
-    NESOASSERT(point.size() == static_cast<std::size_t>(ndim),
-               "Miss-match in point size and mesh dimension.");
-    NESOASSERT(2 == ndim, "Only implemented in 2D.");
-    this->check_valid_local_cell(index);
-    const PetscInt petsc_index = this->map_np_to_petsc.at(index);
-    bool contained = false;
-    const PetscScalar x0 = point.at(0);
-    const PetscScalar x1 = point.at(1);
-
-    const PetscScalar *tmp;
-    PetscScalar *vertices = nullptr;
-    PetscInt num_crossings = 0, num_coords;
-    PetscBool is_dg;
-
-    PETSCCHK(DMPlexGetCellCoordinates(dm, petsc_index, &is_dg, &num_coords,
-                                      &tmp, &vertices));
-    const int num_faces = num_coords / 2;
-    int faces[8];
-    if (num_coords == 8) {
-      // {0, 1, 1, 2, 2, 3, 3, 0};
-      faces[0] = 0;
-      faces[1] = 1;
-      faces[2] = 1;
-      faces[3] = 2;
-      faces[4] = 2;
-      faces[5] = 3;
-      faces[6] = 3;
-      faces[7] = 0;
-    } else {
-      // {0, 1, 1, 2, 2, 0};
-      faces[0] = 0;
-      faces[1] = 1;
-      faces[2] = 1;
-      faces[3] = 2;
-      faces[4] = 2;
-      faces[5] = 0;
-    }
-
-    for (int facex = 0; facex < num_faces; facex++) {
-      REAL xi = vertices[faces[2 * facex + 0] * 2 + 0];
-      REAL yi = vertices[faces[2 * facex + 0] * 2 + 1];
-      REAL xj = vertices[faces[2 * facex + 1] * 2 + 0];
-      REAL yj = vertices[faces[2 * facex + 1] * 2 + 1];
-      // Is the point in a corner
-      if ((x0 == xj) && (x1 == yj)) {
-        num_crossings = 1;
-        break;
-      }
-      if ((yj > x1) != (yi > x1)) {
-        REAL determinate = (x0 - xj) * (yi - yj) - (xi - xj) * (x1 - yj);
-        if (determinate == 0) {
-          // Point is on line
-          num_crossings = 1;
-          break;
-        }
-        if ((determinate < 0) != (yi < yj)) {
-          num_crossings++;
-        }
-      }
-    }
-
-    // odd number of crossings implies the point is contained
-    if ((num_crossings % 2) == 1) {
-      contained = true;
-    };
-    PETSCCHK(DMPlexRestoreCellCoordinates(dm, petsc_index, &is_dg, &num_coords,
-                                          &tmp, &vertices));
-
-    return contained;
-  }
+  bool cell_contains_point(const PetscInt index,
+                           std::vector<PetscScalar> &point);
 
   /**
    * @returns The number of labels in the DMPlex.
    */
-  inline PetscInt get_num_labels() {
-    PetscInt num_labels;
-    PETSCCHK(DMGetNumLabels(this->dm, &num_labels));
-    return num_labels;
-  }
+  PetscInt get_num_labels();
 
   /**
    * @returns The DMLabel for "Face Sets"
    */
-  inline DMLabel get_face_sets_label() {
-    const char *name = "Face Sets";
-    PetscBool has_label;
-    PETSCCHK(DMHasLabel(this->dm, name, &has_label));
-    NESOASSERT(has_label,
-               "The Face Sets label does not exist on this DMPlex. If using "
-               "gmsh, check Physical Lines and Physical Surfaces are set.");
-    DMLabel label;
-    PETSCCHK(DMGetLabel(this->dm, name, &label));
-    return label;
-  }
+  DMLabel get_face_sets_label();
 
   /**
    * @param index Index of label to get name of.
    * @returns The label name corresponding to an index.
    */
-  inline std::string get_label_name(const PetscInt index) {
-    const char *name;
-    PETSCCHK(DMGetLabelName(this->dm, index, &name));
-    return std::string(name);
-  }
+  std::string get_label_name(const PetscInt index);
 
   /**
    * Get boundary points start and end. In 2D the boundary point types are
@@ -807,68 +360,23 @@ public:
    * @param[in, out] start First boundary point.
    * @param[in, out] end Last boundary point plus one.
    */
-  inline void get_boundary_stratum(PetscInt *start, PetscInt *end) {
-    PetscInt depth = this->ndim - 1;
-    PETSCCHK(DMPlexGetDepthStratum(this->dm, depth, start, end));
-  }
+  void get_boundary_stratum(PetscInt *start, PetscInt *end);
 
   /**
    * @returns Map from face sets int label to DMPlex points with that label.
    */
-  inline std::map<PetscInt, std::vector<PetscInt>> get_face_sets() {
-    DMLabel face_sets_label = this->get_face_sets_label();
-    PetscInt points_start, points_end;
-    this->get_boundary_stratum(&points_start, &points_end);
-
-    std::map<PetscInt, std::vector<PetscInt>> map;
-    for (PetscInt px = points_start; px < points_end; px++) {
-      PetscInt value;
-      PETSCCHK(DMLabelGetValue(face_sets_label, px, &value));
-      map[value].push_back(px);
-    }
-
-    return map;
-  }
+  std::map<PetscInt, std::vector<PetscInt>> get_face_sets();
 
   /**
    * Write the DM to a file for visualisation in paraview.
    *
    * @param filename Filename for VTK file.
    */
-  inline void write_vtk(const std::string filename) {
-    PetscViewer viewer;
-    PETSCCHK(PetscViewerCreate(PETSC_COMM_SELF, &viewer));
-    PETSCCHK(PetscViewerSetType(viewer, PETSCVIEWERVTK));
-    PETSCCHK(PetscViewerFileSetMode(viewer, FILE_MODE_WRITE));
-    PETSCCHK(PetscViewerFileSetName(viewer, filename.c_str()));
-    PETSCCHK(DMView(this->dm, viewer));
-    PETSCCHK(PetscViewerDestroy(&viewer));
-  }
-
+  void write_vtk(const std::string filename);
   /**
    * Print to stdout information about the held DMPlex.
    */
-  inline void print() {
-    for (int cx = 0; cx < this->ncells; cx++) {
-      const PetscInt point_index = this->map_np_to_petsc.at(cx);
-      nprint("---------------------------------------------------------------");
-      nprint("Local index:", cx, "point index:", point_index);
-      PetscBool is_dg;
-      PetscInt nc;
-      const PetscScalar *array;
-      PetscScalar *coords;
-      PETSCCHK(DMPlexGetCellCoordinates(dm, cx, &is_dg, &nc, &array, &coords));
-      for (int px = 0; px < nc; px += this->ndim) {
-        std::cout << "\t";
-        for (int dx = 0; dx < this->ndim; dx++) {
-          std::cout << coords[px + dx] << " ";
-        }
-        std::cout << std::endl;
-      }
-      PETSCCHK(
-          DMPlexRestoreCellCoordinates(dm, cx, &is_dg, &nc, &array, &coords));
-    }
-  }
+  void print();
 
   /**
    * Get the volume of a cell in the local mesh.
@@ -876,32 +384,13 @@ public:
    * @param cell Local index of cell.
    * @returns Volume of cell.
    */
-  inline REAL get_cell_volume(const int index) {
-    this->check_valid_local_cell(index);
-    const PetscInt petsc_index = this->map_np_to_petsc.at(index);
-    PetscReal vol;
-    PetscReal centroid[3];
-    PetscReal normal[3];
-    PETSCCHK(DMPlexComputeCellGeometryFVM(this->dm, petsc_index, &vol, centroid,
-                                          normal));
-    return vol;
-  }
+  REAL get_cell_volume(const int index);
 
   /**
    * @returns The total volume of the mesh. Must be called collectively on the
    * communicator.
    */
-  inline REAL get_volume() {
-    if (this->volume < 0.0) {
-      double local_volume = 0.0;
-      for (int cx = 0; cx < this->ncells; cx++) {
-        local_volume += this->get_cell_volume(cx);
-      }
-      MPICHK(MPI_Allreduce(&local_volume, &this->volume, 1, MPI_DOUBLE, MPI_SUM,
-                           this->comm));
-    }
-    return this->volume;
-  }
+  REAL get_volume();
 };
 
 /**
@@ -909,43 +398,10 @@ public:
  *
  * @returns CellDatConst for number of cell vertices and cell vertices.
  */
-inline std::tuple<std::shared_ptr<CellDatConst<int>>,
-                  std::shared_ptr<CellDatConst<REAL>>>
+std::tuple<std::shared_ptr<CellDatConst<int>>,
+           std::shared_ptr<CellDatConst<REAL>>>
 get_cell_vertices_cdc(SYCLTargetSharedPtr sycl_target,
-                      std::shared_ptr<DMPlexHelper> dmh) {
-  std::tuple<std::shared_ptr<CellDatConst<int>>,
-             std::shared_ptr<CellDatConst<REAL>>>
-      d;
-  const int cell_count = dmh->get_cell_count();
-
-  std::vector<std::vector<std::vector<REAL>>> vertices(cell_count);
-
-  std::size_t max_num_vertices = 0;
-  const auto ndim = dmh->ndim;
-  for (int cellx = 0; cellx < cell_count; cellx++) {
-    dmh->get_cell_vertices(cellx, vertices.at(cellx));
-    max_num_vertices = std::max(max_num_vertices, vertices.at(cellx).size());
-  }
-
-  std::get<0>(d) =
-      std::make_shared<CellDatConst<int>>(sycl_target, cell_count, 1, 1);
-  std::get<1>(d) = std::make_shared<CellDatConst<REAL>>(sycl_target, cell_count,
-                                                        max_num_vertices, ndim);
-
-  for (int cellx = 0; cellx < cell_count; cellx++) {
-    const int num_vertices = vertices.at(cellx).size();
-    std::get<0>(d)->set_value(cellx, 0, 0, num_vertices);
-    auto c = std::get<1>(d)->get_cell(cellx);
-    for (int rx = 0; rx < num_vertices; rx++) {
-      for (int cx = 0; cx < ndim; cx++) {
-        c->at(rx, cx) = vertices.at(cellx).at(rx).at(cx);
-      }
-    }
-    std::get<1>(d)->set_cell(cellx, c);
-  }
-
-  return d;
-}
+                      std::shared_ptr<DMPlexHelper> dmh);
 
 } // namespace NESO::Particles::PetscInterface
